GO Project structure
