--- backend/go.mod
@@ -1,16 +1,0 @@
-module github.com/keylime/keylime-mcp/backend
-
-go 1.23.0
-
-toolchain go1.24.8
-
-require (
-	github.com/joho/godotenv v1.5.1
-	github.com/modelcontextprotocol/go-sdk v1.1.0
-)
-
-require (
-	github.com/google/jsonschema-go v0.3.0 // indirect
-	github.com/yosida95/uritemplate/v3 v3.0.2 // indirect
-	golang.org/x/oauth2 v0.30.0 // indirect
-)
