--- compose.yml
@@ -1,41 +1,0 @@
-services:
-  backend:
-    build: ./backend
-    container_name: keylime-mcp-backend
-    network_mode: host
-    env_file:
-      - .env
-    environment:
-      - KEYLIME_CERT_DIR=/certs
-      - RUN_MODE=http
-    volumes:
-      - ${HOME}/.keylime/certs:/certs:ro,z
-    security_opt:
-      - no-new-privileges:true
-
-  frontend:
-    build:
-      context: ./frontend
-      target: builder
-    container_name: keylime-mcp-frontend
-    ports:
-      - "3000:5173"
-    volumes:
-      - ./frontend/src:/app/src
-      - ./frontend/public:/app/public
-      - ./frontend/index.html:/app/index.html
-      - ./frontend/vite.config.ts:/app/vite.config.ts
-      - ./frontend/tsconfig.json:/app/tsconfig.json
-      - ./frontend/tsconfig.app.json:/app/tsconfig.app.json
-    command: pnpm dev --host
-    depends_on:
-      - backend
-    networks:
-      - keylime-network
-    security_opt:
-      - no-new-privileges:true
-
-networks:
-  keylime-network:
-    driver: bridge
-
