--- frontend/nginx.conf
@@ -1,15 +1,0 @@
-server {
-    listen 80;
-    root /usr/share/nginx/html;
-    index index.html;
-
-    location / {
-        try_files $uri $uri/ /index.html;
-    }
-
-    location /api {
-        proxy_pass http://backend:8080;
-        rewrite ^/api/(.*) /$1 break;
-    }
-}
-
